redid config files

+++ lgt
@@ -1,41 +1,57 @@
 #!/usr/bin/env bash
 
-archive=${HOME}/.lgtdb
-tmpfile="/tmp/.${USER}.lgt"
+# TO DO: get list of languages and check if languages are valid
+# TO DO: use authtoken to make requests faster
 
-languages="
-makefile
-c
-cpp
-coffeescript
-css
-go
-haskell
-html
-javascript
-perl
-python
-r
-ruby
-rust
-shell
-"
+lgtdir=${HOME}/.lgt
+archive=$lgtdir/archive
+tmpfile=$lgtdir/tmp
+lanfile=$lgtdir/languages
+tokfile=$lgtdir/token
 
-rm -f $archive
+if [ ! -e $lgtdir ]; then
+  mkdir -p $lgtdir
+fi
 
-touch $archive
+if [ ! -e $lanfile ]; then
+  echo -e 'c\ncpp\nshell' > $lanfile
+fi
+
+if [ ! -e $lgtdir/token ]; then
+  hastoken=false
+else
+  hastoken=true
+fi
+
+if [ ! -e $archive ]; then
+  touch $archive
+fi
+
+
+
+if [ $# -gt 0 ]; then
+  if [ $1 = "token" ]; then
+    if [ $# -lt 2 ]; then
+      echo "Please supply personal authentication token"
+    else
+      echo $2 > $tokfile
+    fi
+  else
+    languages=$1
+  fi
+else
+  languages=$(cat $lanfile | tr '\n' ' ')
+fi
 
 for i in $languages; do
-
   echo $i
   curl -sG https://api.github.com/search/repositories --data-urlencode "sort=stars" --data-urlencode "order=desc" --data-urlencode "q=language:${i}" | jq ".items[0,1,2,3,4,5,6,7,8,9,10] | {description,html_url,language,name}" | sed 's/.*": "//g' | grep -v '{' | grep -v '}' | sed 's/",$//g' | paste - - - - >> $tmpfile
   sleep 6
 done
 
-grep -vFf <(cut -f 3 $archive) $tmpfile | tee -a $archive | awk '{print $0,""}' OFS='\t' | tr '\t' '\n'
-
-rm $tmpfile
 
 
-
-
+if [ -e $tmpfile ]; then
+  grep -vFf <(cut -f 3 $archive) $tmpfile | tee -a $archive | awk '{print $0,""}' OFS='\t' | tr '\t' '\n'
+  rm $tmpfile
+fi
